Lobby & Rooms
Beginning of implementing Multiplayer. Added a response for entering the lobby "Free Channel" so game wouldn't crash or hang when entering the lobby. Added an example room which is currently being used for testing. Currently there's no way to join rooms and no implementation to create rooms from the client.