Refactored (Experimental) & Tools
Experimenting with a change in serialization, it's the same method just presented in a different way to make the offset positions of properties more clear, essentially just refactored code, if this is how I want to continue coding the serialization of packets, then the rest of the code will be refactored too. Added the beginnings of joining a room. A quick python tool is added to analyze packets, updates in the future are expected to add more functionality. Fixed a bug in ZM Stat packet serialization.