Updated Rooms
Rooms are now being read and displayed from a json file. However, in the process, a bug has formed which displays the rooms as in game instead of lobby/waiting room.  There are still many unknown fields and a new field might be required to store what I assume to be the guild of the room master. Refactored some of the code as well.