Match Implementation
Starting a match gets into loading screen, there's a problem with the packets and it loads forever. Pretty confident it's an issue with the Notify message action: 0x23, at bytes 6 and 7 (assuming it's one field.) This is most likely a type of ID generated by the P2P Server. Solving is going to be difficult.