Room Improvements
Items are somewhat fixed, Session ID seems to not need to be accurate, the client will throw an error, but jumping over the error seems completely fine on the official server. "모든 P2P 연결 성공" Message in the logs doesn't appear in the logs by default, bypassing the error sends the next expected packet that was being hung before, however it doesn't fix the infinite "Connecting". Also I need to stop adding my issues in commit messages, and instead open issues on Github.