Improvements & Bugfixes
Significant RE progress in regards to Avatar Lists and a link to Multiplayer crashing based off the Items provided to the Avatars (Things such as matching ServerIDs in user inventory and user Avatar). Some items MUST be in a users inventory such as Default Exo and Fishing Rod Etc, absence of such items triggers packets being sent by the Client which are still "Unknown". As of right now only 1 Client crashes in Multiplayer (improvement from both crashing), supposedly because the second Client is currently using the same account data as the other, this means the second user doesn't own any of the items thus the crashing could be a Anti-Cheat measure? TLDR: Avatars need to be completely implemented before Multiplayer can work. Many bugs are caused by incomplete or invalid data despite not seemingly related.